requirements.txt
pandas==1.4.1
openpyxl==3.1.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Deprecated==1.2.14
idna==3.10
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
requests==2.32.3
typing-extensions==4.12.2
urllib3==2.2.3
wrapt==1.16.0
python-dotenv==1.0.1